Modify ee-add-remove benchmark
Deleted the ee-newlistener benchmark and modified the ee-add-remove
benchmark with the changes.
